refactor: merge TS compiler into single file (#5091)
