+++ cli/js/deno_unstable.ts
@@ -7,7 +7,7 @@ export { linkSync, link } from "./ops/fs/link.ts";
 export { symlinkSync, symlink } from "./ops/fs/symlink.ts";
 export { dir, loadavg, osRelease } from "./ops/os.ts";
 export { openPlugin } from "./ops/plugins.ts";
-export { transpileOnly, compile, bundle } from "./compiler/api.ts";
+export { transpileOnly, compile, bundle } from "./compiler_api.ts";
 export { applySourceMap, formatDiagnostics } from "./ops/errors.ts";
 export { signal, signals, Signal, SignalStream } from "./signals.ts";
 export { setRaw } from "./ops/tty.ts";
