+++ cli/js/ts_global.d.ts
