+++ cli/tests/020_json_modules.ts.out
@@ -1,9 +1,9 @@
 [WILDCARD]
 error: Uncaught TypeError: Cannot resolve extension for "[WILDCARD]config.json" with mediaType "Json".
-    at getExtension ($deno$/compiler/sourcefile.ts:[WILDCARD])
-    at new SourceFile ($deno$/compiler/sourcefile.ts:[WILDCARD])
-    at processImports ($deno$/compiler/imports.ts:[WILDCARD])
-    at async Object.processImports ($deno$/compiler/imports.ts:[WILDCARD])
+    at getExtension ($deno$/compiler.ts:[WILDCARD])
+    at new SourceFile ($deno$/compiler.ts:[WILDCARD])
+    at processImports ($deno$/compiler.ts:[WILDCARD])
+    at async processImports ($deno$/compiler.ts:[WILDCARD])
     at async compile ($deno$/compiler.ts:[WILDCARD])
     at async tsCompilerOnMessage ($deno$/compiler.ts:[WILDCARD])
     at async workerMessageRecvCallback ($deno$/runtime_worker.ts:[WILDCARD])
