+++ cli/tests/error_011_bad_module_specifier.ts.out
@@ -1,9 +1,9 @@
 [WILDCARD]error: Uncaught URIError: relative import path "bad-module.ts" not prefixed with / or ./ or ../ Imported from "[WILDCARD]/error_011_bad_module_specifier.ts"
     at unwrapResponse ($deno$/ops/dispatch_json.ts:[WILDCARD])
     at Object.sendSync ($deno$/ops/dispatch_json.ts:[WILDCARD])
-    at resolveModules ($deno$/compiler/imports.ts:[WILDCARD])
-    at processImports ($deno$/compiler/imports.ts:[WILDCARD])
-    at Object.processImports ($deno$/compiler/imports.ts:[WILDCARD])
+    at resolveModules ($deno$/compiler.ts:[WILDCARD])
+    at processImports ($deno$/compiler.ts:[WILDCARD])
+    at processImports ($deno$/compiler.ts:[WILDCARD])
     at async compile ($deno$/compiler.ts:[WILDCARD])
     at async tsCompilerOnMessage ($deno$/compiler.ts:[WILDCARD])
     at async workerMessageRecvCallback ([WILDCARD]runtime_worker.ts:[WILDCARD])
