+++ cli/tests/error_local_static_import_from_remote.ts.out
@@ -2,8 +2,8 @@
 error: Uncaught PermissionDenied: Remote module are not allowed to statically import local modules. Use dynamic import instead.
     at unwrapResponse ($deno$/ops/dispatch_json.ts:[WILDCARD])
     at Object.sendAsync ($deno$/ops/dispatch_json.ts:[WILDCARD])
-    at async processImports ($deno$/compiler/imports.ts:[WILDCARD])
-    at async Object.processImports ($deno$/compiler/imports.ts:[WILDCARD])
+    at async processImports ($deno$/compiler.ts:[WILDCARD])
+    at async processImports ($deno$/compiler.ts:[WILDCARD])
     at async compile ($deno$/compiler.ts:[WILDCARD])
     at async tsCompilerOnMessage ($deno$/compiler.ts:[WILDCARD])
     at async workerMessageRecvCallback ($deno$/runtime_worker.ts:[WILDCARD])
