+++ cli/tests/error_type_definitions.ts.out
@@ -1,11 +1,10 @@
 [WILDCARD]error: Uncaught URIError: relative import path "baz" not prefixed with / or ./ or ../ Imported from "[WILDCARD]/type_definitions/bar.d.ts"
     at unwrapResponse ($deno$/ops/dispatch_json.ts:[WILDCARD])
     at Object.sendSync ($deno$/ops/dispatch_json.ts:[WILDCARD])
-    at Object.resolveModules ($deno$/ops/compiler.ts:[WILDCARD])
-    at resolveModules ($deno$/compiler/imports.ts:[WILDCARD])
-    at processImports ($deno$/compiler/imports.ts:[WILDCARD])
-    at processImports ($deno$/compiler/imports.ts:[WILDCARD])
-    at async Object.processImports ($deno$/compiler/imports.ts:[WILDCARD])
+    at resolveModules ($deno$/compiler.ts:[WILDCARD])
+    at processImports ($deno$/compiler.ts:[WILDCARD])
+    at processImports ($deno$/compiler.ts:[WILDCARD])
+    at async processImports ($deno$/compiler.ts:[WILDCARD])
     at async compile ($deno$/compiler.ts:[WILDCARD])
     at async tsCompilerOnMessage ($deno$/compiler.ts:[WILDCARD])
     at async workerMessageRecvCallback ([WILDCARD]runtime_worker.ts:[WILDCARD])
